# FastOr: force using bitmap containers