@lemire 2017-10-18
This code looks very similar to both the Java and C code that we have thoroughly benchmark. Furthermore, you indicate that you are getting good results on private benchmarks. That's good enough for me.