@lemire 2017-10-18
I'll issue a release.