@maciej 2017-10-19
@lemire great, thank you!