@lemire 2021-08-07
PR https://github.com/RoaringBitmap/roaring/pull/331 is going to revert this PR.

Further investigations reveals that it could lead to containers containing mere dozens of entries... Spending 8kB on a few entries can boost performance for sure, but there is a significant tradeoff involved.

In the Java or C++ implementation of roaring, we do not ever go for bitmap containers with less than or equal to 4096 entries. It is a constant. I think it is best to stay with the standard format and not to trade memory usage away for performance. The results could be unfortunate.

Further PRs to improve the performance are certainly invited. I have added some checks to help us avoid such deviations in the future.